+++ wazo_auth/plugins/backends/mock.py
@@ -1,29 +1,33 @@
 # -*- coding: utf-8 -*-
-# Copyright 2015-2017 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2018 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0+
 
 from wazo_auth import BaseAuthenticationBackend
 
 
-class BackendMock(BaseAuthenticationBackend):
+class _BaseMockBackend(BaseAuthenticationBackend):
+
+    login, password = 'foo', 'bar'
 
     def get_acls(self, login, args):
-        return ['foo', 'bar', 'auth.#']
+        return self._acls
 
-    def get_ids(self, login, agrs):
-        return 'a-mocked-uuid', None
-
-    def verify_password(self, login, password, args):
-        return login == 'foo' and password == 'bar'
-
-
-class BackendMockWithUUID(BaseAuthenticationBackend):
-
-    def get_acls(self, login, args):
-        return ['foo', 'bar']
-
-    def get_ids(self, login, args):
-        return 'a-mocked-auth-id', 'a-mocked-xivo-user-uuid'
+    def get_metadata(self, login, args):
+        metadata = super(_BaseMockBackend, self).get_metadata(login, args)
+        metadata.update(self._base_metadata)
+        return metadata
 
     def verify_password(self, login, password, args):
-        return login == 'foo' and password == 'bar'
+        return (login, password) == (self.login, self.password)
+
+
+class BackendMock(_BaseMockBackend):
+
+    _base_metadata = dict(auth_id='a-mocked-uuid')
+    _acls = ['foo', 'bar', 'auth.#']
+
+
+class BackendMockWithUUID(_BaseMockBackend):
+
+    _base_metadata = dict(auth_id='a-mocked-auth-id', xivo_user_uuid='a-mocked-xivo-user-uuid')
+    _acls = ['foo', 'bar']