+++ wazo_auth/plugins/backends/tests/test_xivo_service.py
@@ -1,11 +1,11 @@
 # -*- coding: utf-8 -*-
-# Copyright 2015-2017 The Wazo Authors  (see the AUTHORS file)
+# Copyright 2015-2018 The Wazo Authors  (see the AUTHORS file)
 # SPDX-License-Identifier: GPL-3.0+
 
 import unittest
 
 from mock import Mock, patch
-from hamcrest import assert_that, calling, equal_to, raises
+from hamcrest import assert_that, calling, equal_to, has_entries, raises
 
 from wazo_auth.plugins import backends
 from wazo_auth.exceptions import AuthenticationFailedException
@@ -23,6 +23,9 @@ class TestVerifyPassword(unittest.TestCase):
         assert_that(result, equal_to('a_return_value'))
         dao_mock.assert_called_once_with('foo', 'bar')
 
+
+class TestGetAcls(unittest.TestCase):
+
     @patch('wazo_auth.plugins.backends.xivo_service.accesswebservice_dao.get_user_acl',
            Mock(return_value=['confd.#', 'dird.#']))
     def test_that_get_acls_return_acl_for_confd(self):
@@ -32,15 +35,20 @@ class TestVerifyPassword(unittest.TestCase):
 
         assert_that(acls, equal_to(['confd.#', 'dird.#']))
 
+
+class TestGetMetadata(unittest.TestCase):
+
     @patch('wazo_auth.plugins.backends.xivo_service.accesswebservice_dao.get_user_uuid',
            Mock(return_value='534ede0d-9395-445a-8541-96b99e7b16a5'))
-    def test_that_get_ids_returns_the_id_and_None(self):
+    def test_that_get_metadata_returns_the_id_and_None(self):
         backend = backends.XiVOService()
 
-        auth_id, xivo_user_uuid = backend.get_ids('foo', None)
+        result = backend.get_metadata('foo', None)
 
-        assert_that(auth_id, equal_to('534ede0d-9395-445a-8541-96b99e7b16a5'))
-        assert_that(xivo_user_uuid, equal_to(None))
+        assert_that(result, has_entries(
+            auth_id='534ede0d-9395-445a-8541-96b99e7b16a5',
+            xivo_user_uuid=None,
+        ))
 
     @patch('wazo_auth.plugins.backends.xivo_service.accesswebservice_dao.get_user_uuid',
            Mock(side_effect=LookupError))
@@ -48,5 +56,5 @@ class TestVerifyPassword(unittest.TestCase):
         backend = backends.XiVOService()
 
         assert_that(
-            calling(backend.get_ids).with_args('foo', None),
+            calling(backend.get_metadata).with_args('foo', None),
             raises(AuthenticationFailedException))